+++ dynomite-derive/examples/demo.rs
@@ -1,15 +1,29 @@
+#[macro_use]
 extern crate dynomite;
 #[macro_use]
 extern crate dynomite_derive;
+extern crate futures;
 extern crate rusoto_core;
-extern crate rusoto_dynamodb;
+extern crate tokio;
 extern crate uuid;
 
-use rusoto_dynamodb::*;
-use uuid::Uuid;
+use std::sync::Arc;
 
-// for Item trait interface resolution
+use dynamodb::{
+    AttributeDefinition, CreateTableInput, DynamoDb, DynamoDbClient, GetItemInput,
+    KeySchemaElement, ProvisionedThroughput, PutItemInput, ScanInput,
+};
+// dynomite re-exports `rusoto_dynamodb` for convenience
+use dynomite::dynamodb;
+// this enables extension methods on `DynamoDB` clients
+use dynomite::DynamoDbExt;
+// this enables a types to be coersed from attribute maps
+use dynomite::FromAttributes;
+// this enables `Item` methods on types which Item is implemented or derived for
 use dynomite::Item;
+use futures::{Future, Stream};
+use tokio::runtime::Runtime;
+use uuid::Uuid;
 
 #[derive(Item, Debug, Clone)]
 pub struct Book {
@@ -20,8 +34,9 @@ pub struct Book {
 
 // this will create a rust book shelf in your aws account!
 fn main() {
+    let mut rt = Runtime::new().expect("failed to initialize futures runtime");
     // create rusoto client
-    let client = DynamoDbClient::new(Default::default());
+    let client = Arc::new(DynamoDbClient::new(Default::default()));
 
     // create a book table with a single string (S) primary key.
     // if this table does not already exists
@@ -29,24 +44,22 @@ fn main() {
     // it will fail if this table already exists but that's okay,
     // this is just an example :)
     let table_name = "books".to_string();
-    let _ = client
-        .create_table(CreateTableInput {
-            table_name: table_name.clone(),
-            key_schema: vec![KeySchemaElement {
-                attribute_name: "id".into(),
-                key_type: "HASH".into(),
-            }],
-            attribute_definitions: vec![AttributeDefinition {
-                attribute_name: "id".into(),
-                attribute_type: "S".into(),
-            }],
-            provisioned_throughput: ProvisionedThroughput {
-                read_capacity_units: 1,
-                write_capacity_units: 1,
-            },
-            ..Default::default()
-        })
-        .sync();
+    let _ = rt.block_on(client.create_table(CreateTableInput {
+        table_name: table_name.clone(),
+        key_schema: vec![KeySchemaElement {
+            attribute_name: "id".into(),
+            key_type: "HASH".into(),
+        }],
+        attribute_definitions: vec![AttributeDefinition {
+            attribute_name: "id".into(),
+            attribute_type: "S".into(),
+        }],
+        provisioned_throughput: ProvisionedThroughput {
+            read_capacity_units: 1,
+            write_capacity_units: 1,
+        },
+        ..CreateTableInput::default()
+    }));
 
     let book = Book {
         id: Uuid::new_v4(),
@@ -55,29 +68,62 @@ fn main() {
 
     // print the key for this book
     // requires bringing `dynomite::Item` into scope
-    println!("key {:#?}", book.key());
+    println!("book.key() {:#?}", book.key());
 
     // add a book to the shelf
     println!(
-        "{:#?}",
-        client
-            .put_item(PutItemInput {
-                table_name: table_name.clone(),
-                item: book.clone().into(), // convert book into it's attribute representation
-                ..Default::default()
-            })
-            .sync()
+        "put_item() result {:#?}",
+        rt.block_on(client.put_item(PutItemInput {
+            table_name: table_name.clone(),
+            item: book.clone().into(), // <= convert book into it's attribute map representation
+            ..PutItemInput::default()
+        }))
     );
 
-    // get the book by it's application generated key
     println!(
-        "{:#?}",
-        client
-            .get_item(GetItemInput {
+        "put_item() result {:#?}",
+        rt.block_on(
+            client.put_item(PutItemInput {
                 table_name: table_name.clone(),
-                key: book.clone().key(), // get a book by key
-                ..Default::default()
+                // convert book into it's attribute map representation
+                item: Book {
+                    id: Uuid::new_v4(),
+                    title: "rust and beyond".into(),
+                }.into(),
+                ..PutItemInput::default()
             })
-            .sync()
+        )
+    );
+
+    // scan through all pages of results in the books table for books who's title is "rust"
+    println!(
+        "scan result {:#?}",
+        rt.block_on(
+            client
+                .clone()
+                .stream_scan(ScanInput {
+                    limit: Some(1), // to demonstrate we're getting through more than one page
+                    table_name: table_name.clone(),
+                    filter_expression: Some("title = :title".into()),
+                    expression_attribute_values: Some(attr_map!(
+                        ":title" => "rust".to_string()
+                    )),
+                    ..ScanInput::default()
+                })
+                .for_each(|item| Ok(println!("stream_scan() item {:#?}", Book::from_attrs(item)))) // attempt to convert a attribute map to a book type
+        ),
+    );
+
+    // get the "rust' book by the Book type's generated key
+    println!(
+        "get_item() result {:#?}",
+        rt.block_on(
+            client
+                .get_item(GetItemInput {
+                    table_name: table_name.clone(),
+                    key: book.clone().key(), // get a book by key
+                    ..GetItemInput::default()
+                }).map(|result| result.item.map(|item| Book::from_attrs(item))) // attempt to convert a attribute map to a book type
+        )
     );
 }